requirements.txt
python-dotenv==1.0.0
fastapi==0.115.7
uvicorn==0.34.0
pydantic-settings==2.7.1
orjson==3.10.15
pydantic==2.10.5
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
alembic==1.14.1
psycopg2-binary==2.9.5
IMAPClient==3.0.1
passlib==1.7.4
python-jose==3.3.0